+++ tests/testthat/developer.api.autodesk.com/oss/v2/buckets/mybucket-DELETE.R
@@ -1,0 +1,11 @@
+structure(
+  list(
+    method     = "DELETE",
+    url        = "https://developer.api.autodesk.com/oss/v2/buckets/mybucket",
+    status_code = 200L,
+    headers    = structure(list(`content-type` = "application/json; charset=utf-8"),
+                           class = "httr2_headers"),
+    body       = charToRaw("{}")
+  ),
+  class = "httr2_response"
+)